Refactoring

+++ dataset/data/prepare_data.py
@@ -6,12 +6,6 @@ from time import time
 def get_args():
     parser = argparse.ArgumentParser()
     parser.add_argument(
-        "--dataset_path",
-        type=str,
-        required=True,
-        help="Path to the shards."
-    )
-    parser.add_argument(
         "--working_dir",
         type=str,
         required=True,
@@ -56,11 +50,11 @@ def get_args():
         type=str,
         required=True,
     )
-    parser.add_argument(
-        "--do_lower_case",
-        type=int,
-        required=True,
-    )
+    # parser.add_argument(
+    #     "--do_lower_case",
+    #     type=bool,
+    #     required=True,
+    # )
     parser.add_argument(
         "--masked_lm_prob",
         type=float,
@@ -94,17 +88,9 @@ def main():
     if not os.path.exists(working_dir):
         os.mkdir(working_dir)
 
-    # 1. Create new dataset from LexicalScores annotated dataset
-    command = f"python3 create_new_dataset.py " \
-              f"--num_workers {args.num_workers} " \
-              f"--dir {args.dataset_path} " \
-              f"-o {working_dir}/1_Dropped/ "
-    os.system(command)
-
-
     # 2. Create n partitions of raw huggingface wikipedia (we need this in case of limited RAM)
     command = f"python3 process_raw_to_shardable.py " \
-              f"--dir {working_dir}/1_Dropped/ " \
+              f"--dir {working_dir}/1_Dataset/ " \
               f"-o {working_dir}/2_Split/ " \
               f"--splits {args.splits} "
     os.system(command)
@@ -134,23 +120,26 @@ def main():
     os.system(command)
     shutil.rmtree(f"{working_dir}/3_Shards/")
 
-    # 7. Generate Samples
-    command = f"python ../generate_samples.py " \
-              f"--dir {working_dir}/4_MergedShards/ " \
-              f"-o {working_dir}/5_MaskedSamples/ " \
-              f"--dup_factor {args.dup_factor} " \
-              f"--seed {args.seed} " \
-              f"--vocab_file {args.vocab_file} " \
-              f"--do_lower_case {args.do_lower_case}  " \
-              f"--masked_lm_prob {args.masked_lm_prob} " \
-              f"--max_seq_length {args.max_seq_length} " \
-              f"--model_name {args.model_name} " \
-              f"--max_predictions_per_seq {args.max_predictions_per_seq} " \
-              f"--n_processes {args.num_workers} "
+    if True:
+        # 7. Generate Samples
+        command = f"python ../generate_samples.py " \
+                  f"--dir {working_dir}/4_MergedShards/ " \
+                  f"-o {working_dir}/5_MaskedSamples/ " \
+                  f"--dup_factor {args.dup_factor} " \
+                  f"--seed {args.seed} " \
+                  f"--vocab_file {args.vocab_file} " \
+                  f"--masked_lm_prob {args.masked_lm_prob} " \
+                  f"--max_seq_length {args.max_seq_length} " \
+                  f"--model_name {args.model_name} " \
+                  f"--max_predictions_per_seq {args.max_predictions_per_seq} " \
+                  f"--n_processes {args.num_workers} "
 
-    os.system(command)
-    shutil.rmtree(f"{working_dir}/4_MergedShards/")
+        os.system(command)
+        shutil.rmtree(f"{working_dir}/4_MergedShards/")
+    
+
     end = time()
+
 
     print("\n" * 5)
     print(f"Finishing processing data within {round((((end-start)/60)/60), 2)} hours.")
